Drop support for TensorFlow 1.13
