+++ setup.py
@@ -24,8 +24,8 @@ setup(
         "dataclasses ; python_version<'3.7'",
     ],
     extras_require={
-        "tensorflow": ["tensorflow>=1.13.1"],
-        "tensorflow_gpu": ["tensorflow-gpu>=1.13.1"],
+        "tensorflow": ["tensorflow>=1.14.0"],
+        "tensorflow_gpu": ["tensorflow-gpu>=1.14.0"],
         "test": [
             "absl-py==0.8.1",
             "pytest==5.2.2",
